move provider types and utils
